Convert prices, budget, etc from int to double

+++ maven/src/main/java/plushie_tycoon/serverService/config/Initials.java
@@ -9,8 +9,8 @@ import java.util.HashMap;
 public class Initials {
     private static int numResources = Resources.values().length;
     public static HashMap<BaseObjects, Integer> quantities = new HashMap<>();
-    public static HashMap<BaseObjects, Integer> prices = new HashMap<>();
-    public static int budget;
+    public static HashMap<BaseObjects, Double> prices = new HashMap<>();
+    public static double budget;
     public static int time;
 
     static {
@@ -23,15 +23,15 @@ public class Initials {
         quantities.put(Products.BETA, 11);
         quantities.put(Products.CHAMA, 12);
 
-        prices.put(Resources.CLOTH, 10);
-        prices.put(Resources.STUFFING, 20);
-        prices.put(Resources.ACCESSORY, 18);
-        prices.put(Resources.PACKAGING, 12);
+        prices.put(Resources.CLOTH, 10.);
+        prices.put(Resources.STUFFING, 20.);
+        prices.put(Resources.ACCESSORY, 18.);
+        prices.put(Resources.PACKAGING, 12.);
 
-        prices.put(Products.AISHA, 102);
-        prices.put(Products.BETA, 103);
-        prices.put(Products.CHAMA, 104);
-        budget = 1000002;
+        prices.put(Products.AISHA, 102.);
+        prices.put(Products.BETA, 103.);
+        prices.put(Products.CHAMA, 104.);
+        budget = 1000002.;
         time = 12;
     }
 
@@ -41,7 +41,7 @@ public class Initials {
     public static int getQuantity(BaseObjects baseObject){
         return quantities.get(baseObject);
     }
-    public static int getPrices(BaseObjects baseObject){
+    public static double getPrices(BaseObjects baseObject){
         return prices.get(baseObject);
     }
 }
